Pulumi running Ansible playbook. Kind of a kludge:
There is no integration between these tools. To "run" Ansible
you need to just execute a raw command. This is not part of Pulumis'
diff process so I have to force trigger it to run the Ansible playbook everytime.

I really expected this to be less hacky but this seems like the only solution atm.
I guess I should have used containers